+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/bohrsty/agenda-display/compare/v0.1.0...HEAD)
 ### Added
+- interval to refresh agenda every "timeout minutes"
 - password field to calendar setting
 - color picker
 - colors in environment, config and global state
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - api endpoint for agenda (data mocked)
 
 ### Changed
+- refresh button to refresh agenda manually
 - api endpoint to encrypt password in redis
 - optimized error handling using api calls in ui
 - caldav connection form to add color